+++ lib/spout/templates/test/dictionary_test.rb
@@ -1,0 +1,10 @@
+require 'test_helper'
+
+class DictionaryTest < Test::Unit::TestCase
+  include Spout::Tests
+
+  # You may add additional tests here
+  # test "the truth" do
+  #   assert true
+  # end
+end